Revert "Bug 1797788: add openshift-tests as a future delegation binary"
