--- images/tests-base/OWNERS
@@ -1,5 +1,0 @@
-reviewers:
-  - smarterclayton
-  - sdodson
-approvers:
-  - smarterclayton
