+++ images/tests/Dockerfile.rhel
@@ -1,15 +1,11 @@
 FROM registry.svc.ci.openshift.org/ocp/builder:golang-1.12 AS builder
 WORKDIR /go/src/github.com/openshift/origin
 COPY . .
-RUN make build WHAT=cmd/openshift-tests-kubernetes; \
-    mkdir -p /tmp/build; \
-    cp /go/src/github.com/openshift/origin/_output/local/bin/linux/$(go env GOARCH)/openshift-tests-kubernetes /tmp/build/openshift-tests-kubernetes
 RUN make build WHAT=cmd/openshift-tests; \
     mkdir -p /tmp/build; \
     cp /go/src/github.com/openshift/origin/_output/local/bin/linux/$(go env GOARCH)/openshift-tests /tmp/build/openshift-tests
 
 FROM registry.svc.ci.openshift.org/ocp/4.2:cli
-COPY --from=builder /tmp/build/openshift-tests-kubernetes /usr/bin/
 COPY --from=builder /tmp/build/openshift-tests /usr/bin/
 RUN yum install --setopt=tsflags=nodocs -y git gzip util-linux && yum clean all && rm -rf /var/cache/yum/* && \
     git config --system user.name test && \
